% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_outline.R
\name{complete_outline}
\alias{complete_outline}
\title{Extract a border segment and join to the ends of an incomplete polygon.}
\usage{
complete_outline(border, overlay)
}
\arguments{
\item{border}{A polygon of lat, lon that outlines an area.}

\item{overlay}{An incomplete polygon of lat, lon to be completed with an extraction from the border.}
}
\value{
A polygon with lat, lon.
}
\description{
Extract a border segment and join to the ends of an incomplete polygon.
}
\examples{
finished_overlay <- complete_outline(peru, overlay)
}
