% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corner2poly.R
\name{corner2poly}
\alias{corner2poly}
\title{Create bounding box coordinates from ll, ur coordinates.}
\usage{
corner2poly(corners)
}
\arguments{
\item{corners}{A dataframe with ll and ur lat, lon.}
}
\value{
A dataframe with five rows of lat, lon.
}
\description{
Create bounding box coordinates from ll, ur coordinates.
}
\examples{
two_points <- data.frame(lon=c(100,110),lat=c(20,25))
bb <- corner2poly(two_points)
}
