% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closest.R
\name{find_closest}
\alias{find_closest}
\title{Find the row number of a border polygon closest to a point.}
\usage{
find_closest(border, point)
}
\arguments{
\item{border}{A dataframe with lat, lon data describing the border outline.}

\item{point}{A pair of lat, lon values.}
}
\value{
The row number of the border dataframe closest to the point.
}
\description{
Find the row number of a border polygon closest to a point.
}
\examples{
nearby <- data.frame(lat = 15, lon = -130)
row_number <- find_closest(boundary, nearby)
}
