% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmap_bb.R
\name{gmap_bb}
\alias{gmap_bb}
\title{Extract the bounding box from a data frame created by site_google_basemap function.}
\usage{
gmap_bb(gmap)
}
\arguments{
\item{gmap}{A dataframe created with the site_google_basemap function.}
}
\value{
A rectangular polygon that is the same as the gmap outline.
}
\description{
Extract the bounding box from a data frame created by site_google_basemap function.
}
\examples{
overlay_poly <- gmap_bb(data_map)
}
