% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawaii_map_theme.R
\name{hawaii_map_theme}
\alias{hawaii_map_theme}
\title{Format without gridlines, color ocean, surround with a line.}
\usage{
hawaii_map_theme(border_color = "black", ocean_color = "lightsteelblue1")
}
\arguments{
\item{border_color}{A color code or name for the map outline.}

\item{ocean_color}{A color code or name for the ocean color.}
}
\value{
code for the theme to use with ggplot.
}
\description{
Format without gridlines, color ocean, surround with a line.
}
\examples{
hawaii_map_theme(ocean_color = "blue")
}
