% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml_polygon.R
\name{kml_polygon}
\alias{kml_polygon}
\title{Convert a Google Earth Pro polygon to geographic coordinates (lat, lon).}
\usage{
kml_polygon(overlay_kml)
}
\arguments{
\item{overlay_kml}{The polygon from Google Earth Pro.}
}
\value{
A dataframe with columns for lat, lon.
}
\description{
Convert a Google Earth Pro polygon to geographic coordinates (lat, lon).
}
\examples{
overlay <- kml_polygon(kml_object)
}
