% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plain_background_theme.R
\name{plain_background_theme}
\alias{plain_background_theme}
\title{Format plot without gridlines or axis annotation.}
\usage{
plain_background_theme(x = NULL)
}
\arguments{
\item{x}{Not used.}
}
\value{
code for the theme to use with ggplot.
}
\description{
Format plot without gridlines or axis annotation.
}
\examples{
plain_background_theme()
}
