% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_plot.R
\name{region_plot}
\alias{region_plot}
\title{Create a map from borders extracted from map_data.}
\usage{
region_plot(
  region,
  fill = "darkseagreen3",
  color = "black",
  linewidth = 0.5,
  coord = 1.3,
  mytheme = NULL
)
}
\arguments{
\item{region}{map_data extract.}

\item{fill}{The color for the interior (in quotes).}

\item{color}{The color for the outline (in quotes).}

\item{linewidth}{Size of the outline line.}

\item{coord}{A scale factor to adjust lat,lon distortion.}

\item{mytheme}{Name of a theme.}
}
\value{
A ggplot-compatible map.
}
\description{
Create a map from borders extracted from map_data.
}
\examples{
region_plot(peru)
region_plot(spain_france, fill="brown", linewidth=0.3)
region_plot(hawaii, coord=1.1)
}
