% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTM_convert.R
\name{UTM_convert}
\alias{UTM_convert}
\title{Convert a dataframe of UTM coordinates to lat, lon coordinates.}
\usage{
UTM_convert(location, zone = 4, hemisphere = "N", km = FALSE)
}
\arguments{
\item{location}{The set of UTM coordinates.}

\item{zone}{A number indicating the UTM section.}

\item{hemisphere}{Whether "N" or "S".}

\item{km}{If the conversion should be km (FALSE = degrees)}
}
\value{
A data frame with lat, lon values
}
\description{
Convert a dataframe of UTM coordinates to lat, lon coordinates.
}
\examples{
niihau_deg <- UTM_convert(niihau_shp,zone=4,hemisphere="N",km=FALSE)
}
